Move httpClient to a separate file
